Move initial set of properties to conf
