+++ src/base/base-mixin.ts
@@ -24,6 +24,7 @@ import {
     ValueAccessor
 } from '../core/types';
 import {IChartGroup} from '../core/chart-group-types';
+import {IBaseConf} from './i-base-conf';
 
 const _defaultFilterHandler = (dimension: MinimalCFDimension, filters) => {
     if (filters.length === 0) {
@@ -83,6 +84,8 @@ const _defaultResetFilterHandler = filters => [];
  * @mixin BaseMixin
  */
 export class BaseMixin {
+    protected _conf:IBaseConf;
+
     // tslint:disable-next-line:variable-name
     private __dcFlag__: string;
     private _dimension: MinimalCFDimension;
@@ -91,10 +94,8 @@ export class BaseMixin {
     private _root: Selection<Element, any, any, any>; // Do not assume much, allow any HTML or SVG element
     private _svg: Selection<SVGElement, any, any, any>; // from d3-selection
     private _isChild: boolean;
-    private _minWidth: number;
     private _defaultWidthCalc: (element) => number;
     private _widthCalc: (element) => number;
-    private _minHeight: number;
     private _defaultHeightCalc: (element) => number;
     private _heightCalc: (element) => number;
     private _width: number;
@@ -129,6 +130,11 @@ export class BaseMixin {
     constructor () {
         this.__dcFlag__ = uniqueId().toString();
 
+        this._conf = {
+            minWidth: 200,
+            minHeight: 200,
+        }
+
         this._dimension = undefined;
         this._group = undefined;
 
@@ -137,17 +143,15 @@ export class BaseMixin {
         this._svg = undefined;
         this._isChild = undefined;
 
-        this._minWidth = 200;
         this._defaultWidthCalc = element => {
             const width = element && element.getBoundingClientRect && element.getBoundingClientRect().width;
-            return (width && width > this._minWidth) ? width : this._minWidth;
+            return (width && width > this._conf.minWidth) ? width : this._conf.minWidth;
         };
         this._widthCalc = this._defaultWidthCalc;
 
-        this._minHeight = 200;
         this._defaultHeightCalc = element => {
             const height = element && element.getBoundingClientRect && element.getBoundingClientRect().height;
-            return (height && height > this._minHeight) ? height : this._minHeight;
+            return (height && height > this._conf.minHeight) ? height : this._conf.minHeight;
         };
         this._heightCalc = this._defaultHeightCalc;
         this._width = undefined;
@@ -199,6 +203,10 @@ export class BaseMixin {
         this._resetFilterHandler = _defaultResetFilterHandler;
     }
 
+    public configure (conf: IBaseConf) {
+        this._conf = {...(this._conf), ...conf}
+    }
+    
     /**
      * Set or get the height attribute of a chart. The height is applied to the SVGElement generated by
      * the chart when rendered (or re-rendered). If a value is given, then it will be used to calculate
@@ -264,40 +272,6 @@ export class BaseMixin {
         }
         this._widthCalc = width ? (typeof width === 'function' ? width : () => width) : this._defaultWidthCalc;
         this._width = undefined;
-        return this;
-    }
-
-    /**
-     * Set or get the minimum width attribute of a chart. This only has effect when used with the default
-     * {@link BaseMixin#width width} function.
-     * @see {@link BaseMixin#width width}
-     * @param {Number} [minWidth=200]
-     * @returns {Number|BaseMixin}
-     */
-    public minWidth (): number;
-    public minWidth (minWidth: number): this;
-    public minWidth (minWidth?) {
-        if (!arguments.length) {
-            return this._minWidth;
-        }
-        this._minWidth = minWidth;
-        return this;
-    }
-
-    /**
-     * Set or get the minimum height attribute of a chart. This only has effect when used with the default
-     * {@link BaseMixin#height height} function.
-     * @see {@link BaseMixin#height height}
-     * @param {Number} [minHeight=200]
-     * @returns {Number|BaseMixin}
-     */
-    public minHeight (): number;
-    public minHeight (minHeight: number): this;
-    public minHeight (minHeight?) {
-        if (!arguments.length) {
-            return this._minHeight;
-        }
-        this._minHeight = minHeight;
         return this;
     }
