+++ src/base/i-base-conf.ts
@@ -1,0 +1,4 @@
+export interface IBaseConf {
+    minWidth?: number;
+    minHeight?: number;
+}
